backend/requirements.txt
fastapi==0.115.0
uvicorn==0.38.0
python-docx==1.1.2
httpx==0.28.1
pydantic==2.10.0